fix: unblock mobile ci tests and max-lines ratchet
